+++ packages/apollo-cache-inmemory/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### vNEXT
 - Added `getCacheKey` function to cacheResolver context [#2998](https://github.com/apollographql/apollo-client/pull/2998)
+- Changed `cacheResolvers` to `cacheRedirects`, added deprecation warning [#3001](https://github.com/apollographql/apollo-client/pull/3001)
 
 ### 1.1.8
 - dependency updates